Renderer & IvyApp

+++ IvyEngine/src/Ivy/Core/HashTable.cpp
@@ -1,7 +1,7 @@
 #include "ivypch.h"
 #include "HashTable.h"
-#include "String.h"
+
 
 namespace Ivy {
-	template class HashTable<uint16_t, String>;
+	template class HashTable<uint32_t, String>;
 }
